get variable star size working, add constants file for high-level game settings

+++ Assets/Scripts/GameEntities/PlanetScript.cs
@@ -6,9 +6,9 @@ public class PlanetScript : MonoBehaviour
 {
 
 
-    public float sizeRadius;
+    public int sizeRadius;
     public int orbitRadius;
-    public float orbitSpeed;
+    public int orbitSpeed;
 
     public float pctWater;
     public float pctGases;
